Move classes to a new src/classes/ directory
To clean up the root
src/ folder.
