+++ src/classes/Warning.ts
